Revert "Introduce formation and squad engines"
