+++ src/engines/worldEngine.js
@@ -1,34 +1,22 @@
 /**
- * \uC6D0\uB4DC \uD130\uB110\uC81C \uC804\uB7B5 \uAC8C\uC784 \uC5D4\uC9C4
- * \uC6D0\uB4DC\uB9F5 \uD0ED \uD751\uBCF5 \uC5ED\uD560\uC744 \uC2E4\uD589\uD569\uB2C8\uB2E4.
+ * \uC804\uB7B5 \uD30C\uD2B8 \uACFC\uC815.
+ * \uC6D0\uB4DC\uB9F5\uC758 \uBAA8\uB4E0 \uAC83\uC744 \uAD8C\uD55C\uD558\uB294 '\uD130\uB110\uC81C \uC804\uB7B5 \uAC8C\uC784 \uC5D4\uC9C4'\uC785\uB2C8\uB2E4.
  */
 export class WorldEngine {
     constructor(game) {
         this.game = game;
+        // \uC6D0\uB4DC\uB9F5\uC740 \uC790\uC2E0\uB9CC\uC758 \uAC00\uB4E4\uD55C GridManager\uC640 TurnManager\uB97C \uAC16\uC2B5\uB2C8\uB2E4.
         this.gridManager = game.gridManager;
         this.turnManager = game.turnManager;
-        this.monsters = [];
     }
 
     update() {
-        const encounteredEnemy = this.checkForEncounters();
-        if (encounteredEnemy) {
-            const terrain = this.getTerrainAt(this.game.player.tileX, this.game.player.tileY);
-            const encounterContext = { enemyUnit: encounteredEnemy, terrain };
-            this.game.setupFormation(encounterContext);
-        }
+        // \uC6D0\uB4DC\uB9F5\uC758 \uD130\uB110 \uB85C\uC9C1 \uCC98\uB9AC
+        // e.g., this.turnManager.update();
     }
 
     render(context) {
-        // TODO: \uC6D0\uB4DC\uB9F5 \uADF8\uB9AC\uAE30
-    }
-
-    checkForEncounters() {
-        const player = this.game.player;
-        return this.monsters.find(m => !m.isDefeated && Math.abs(m.x - player.x) < 30 && Math.abs(m.y - player.y) < 30) || null;
-    }
-
-    getTerrainAt(x, y) {
-        return 'plains';
+        // \uC6D0\uB4DC\uB9F5 \uADF8\uB9AC\uAE30
+        // e.g., this.gridManager.render();
     }
 }
